% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumx.R
\name{sumx}
\alias{sumx}
\title{sumx function}
\usage{
sumx(a, b)
}
\arguments{
\item{a}{first integer parameter}

\item{b}{second integerparameter}
}
\description{
This function provides you with sum of two integer parameters converted to hex string.
}
\examples{
library(pythonTestPackage)
sumx(1,2)
}

